requirements.txt
tgtg
python-telegram-bot